# frontend/openapi/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "PharmaFind API",
      "description": "PharmaFind API Documentation",
      "contact": {
        "name": "OUSSAMA EL HADRAMI",
        "email": "elhadrami.oussama@ensam-casa.ma"
      },
      "version": "1.0"
    },
    "servers": [
      {
        "url": "http://localhost:8080",
        "description": "Local server"
      }
    ],
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "paths": {
      "/products": {
        "get": {
          "tags": [
            "product-controller"
          ],
          "operationId": "getProducts",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 0
              }
            },
            {
              "name": "size",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 10
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PageResponseProductResponse"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "product-controller"
          ],
          "operationId": "updateProduct",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProductResponse"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "product-controller"
          ],
          "operationId": "saveProduct",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "required": [
                    "images",
                    "product"
                  ],
                  "type": "object",
                  "properties": {
                    "product": {
                      "$ref": "#/components/schemas/ProductRequest"
                    },
                    "images": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "format": "binary"
                      }
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProductResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/pharmacies": {
        "get": {
          "tags": [
            "pharmacy-controller"
          ],
          "operationId": "getPharmacies",
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 0
              }
            },
            {
              "name": "size",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 10
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PageResponsePharmacyResponse"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "pharmacy-controller"
          ],
          "operationId": "savePharmacy",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PharmacyRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PharmacyResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/pharmacies/{id}/products": {
        "get": {
          "tags": [
            "pharmacy-controller"
          ],
          "operationId": "getProductsByPharmacy",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "page",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 0
              }
            },
            {
              "name": "size",
              "in": "query",
              "required": false,
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 10
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PageResponseProductResponse"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "pharmacy-controller"
          ],
          "operationId": "saveProduct_1",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "schema": {
                  "required": [
                    "images",
                    "product"
                  ],
                  "type": "object",
                  "properties": {
                    "product": {
                      "$ref": "#/components/schemas/ProductCreateDTO"
                    },
                    "images": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "format": "binary"
                      }
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProductResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/auth/register": {
        "post": {
          "tags": [
            "Authentication"
          ],
          "operationId": "register",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "202": {
              "description": "Accepted",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "/auth/login": {
        "post": {
          "tags": [
            "Authentication"
          ],
          "operationId": "login",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginRequest"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/AuthenticationResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/products/{id}": {
        "get": {
          "tags": [
            "product-controller"
          ],
          "operationId": "getProduct",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ProductResponse"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "product-controller"
          ],
          "operationId": "deleteProduct",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/pharmacies/{id}": {
        "get": {
          "tags": [
            "pharmacy-controller"
          ],
          "operationId": "getPharmacy",
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PharmacyResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/auth/activate": {
        "get": {
          "tags": [
            "Authentication"
          ],
          "operationId": "activate",
          "parameters": [
            {
              "name": "token",
              "in": "query",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/AuthenticationResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "ImageDTO": {
          "type": "object",
          "properties": {
            "imageUrl": {
              "type": "string"
            }
          }
        },
        "ProductRequest": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "price": {
              "type": "number",
              "format": "double"
            },
            "images": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ImageDTO"
              }
            }
          }
        },
        "ProductResponse": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "name": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "price": {
              "type": "number",
              "format": "double"
            },
            "images": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ImageDTO"
              }
            }
          }
        },
        "AddressDTO": {
          "type": "object",
          "properties": {
            "street": {
              "type": "string"
            },
            "city": {
              "type": "string"
            },
            "postalCode": {
              "type": "string"
            }
          }
        },
        "PharmacyRequest": {
          "required": [
            "name",
            "telephone"
          ],
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "telephone": {
              "type": "string"
            },
            "address": {
              "$ref": "#/components/schemas/AddressDTO"
            }
          }
        },
        "PharmacyResponse": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "telephone": {
              "type": "string"
            },
            "address": {
              "$ref": "#/components/schemas/AddressDTO"
            }
          }
        },
        "ProductCreateDTO": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            },
            "description": {
              "type": "string"
            },
            "quantity": {
              "type": "integer",
              "format": "int32"
            },
            "price": {
              "type": "number",
              "format": "double"
            }
          }
        },
        "RegistrationRequest": {
          "required": [
            "email",
            "firstName",
            "lastName",
            "password"
          ],
          "type": "object",
          "properties": {
            "firstName": {
              "type": "string"
            },
            "lastName": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "password": {
              "maxLength": 20,
              "minLength": 6,
              "type": "string"
            },
            "role": {
              "type": "string"
            }
          }
        },
        "LoginRequest": {
          "required": [
            "email",
            "password"
          ],
          "type": "object",
          "properties": {
            "email": {
              "type": "string"
            },
            "password": {
              "maxLength": 20,
              "minLength": 6,
              "type": "string"
            }
          }
        },
        "AuthenticationResponse": {
          "type": "object",
          "properties": {
            "user": {
              "$ref": "#/components/schemas/UserDTO"
            },
            "token": {
              "type": "string"
            }
          }
        },
        "RoleDTO": {
          "type": "object",
          "properties": {
            "name": {
              "type": "string"
            }
          }
        },
        "UserDTO": {
          "type": "object",
          "properties": {
            "firstName": {
              "type": "string"
            },
            "lastName": {
              "type": "string"
            },
            "email": {
              "type": "string"
            },
            "role": {
              "$ref": "#/components/schemas/RoleDTO"
            }
          }
        },
        "PageResponseProductResponse": {
          "type": "object",
          "properties": {
            "content": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ProductResponse"
              }
            },
            "number": {
              "type": "integer",
              "format": "int32"
            },
            "size": {
              "type": "integer",
              "format": "int32"
            },
            "totalElements": {
              "type": "integer",
              "format": "int64"
            },
            "totalPages": {
              "type": "integer",
              "format": "int32"
            },
            "first": {
              "type": "boolean"
            },
            "last": {
              "type": "boolean"
            }
          }
        },
        "PageResponsePharmacyResponse": {
          "type": "object",
          "properties": {
            "content": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/PharmacyResponse"
              }
            },
            "number": {
              "type": "integer",
              "format": "int32"
            },
            "size": {
              "type": "integer",
              "format": "int32"
            },
            "totalElements": {
              "type": "integer",
              "format": "int64"
            },
            "totalPages": {
              "type": "integer",
              "format": "int32"
            },
            "first": {
              "type": "boolean"
            },
            "last": {
              "type": "boolean"
            }
          }
        }
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "description": "JWT Token",
          "in": "header",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    }
  }